final changes for apidoc refactoring
