+++ extensions/apidoc/templates/bootstrap/ApiRenderer.php
@@ -27,28 +27,10 @@ use yii\helpers\Html;
  */
 class ApiRenderer extends \yii\apidoc\templates\html\ApiRenderer
 {
+	use RendererTrait;
+
 	public $layout = '@yii/apidoc/templates/bootstrap/layouts/api.php';
 	public $indexView = '@yii/apidoc/templates/bootstrap/views/index.php';
-
-	public $extensions = [
-		'apidoc',
-		'authclient',
-		'bootstrap',
-		'codeception',
-		'composer',
-		'debug',
-		'elasticsearch',
-		'faker',
-		'gii',
-		'imagine',
-		'jui',
-		'mongodb',
-		'redis',
-		'smarty',
-		'sphinx',
-		'swiftmailer',
-		'twig',
-	];
 
 	/**
 	 * @inheritdoc
@@ -105,57 +87,5 @@ class ApiRenderer extends \yii\apidoc\templates\html\ApiRenderer
 		if ($this->controller !== null) {
 			$this->controller->stdout('done.' . PHP_EOL, Console::FG_GREEN);
 		}
-	}
-
-	public function getNavTypes($type, $types)
-	{
-		if ($type === null) {
-			return $types;
-		}
-		$extensions = $this->extensions;
-		$navClasses = 'app';
-		if (isset($type)) {
-			if ($type->name == 'Yii') {
-				$navClasses = 'yii';
-			} elseif (strncmp($type->name, 'yii\\', 4) == 0) {
-				$subName = substr($type->name, 4);
-				if (($pos = strpos($subName, '\\')) !== false) {
-					$subNamespace = substr($subName, 0, $pos);
-					if (in_array($subNamespace, $extensions)) {
-						$navClasses = $subNamespace;
-					} else {
-						$navClasses = 'yii';
-					}
-				}
-			}
-		}
-		return $this->filterTypes($types, $navClasses);
-	}
-
-	protected function filterTypes($types, $navClasses)
-	{
-		switch ($navClasses)
-		{
-			case 'app':
-				$types = array_filter($types, function($val) {
-					return strncmp($val->name, 'yii\\', 4) !== 0;
-				});
-				break;
-			case 'yii':
-				$self = $this;
-				$types = array_filter($types, function($val) use ($self) {
-					if (strlen($val->name) < 5) {
-						return false;
-					}
-					$subName = substr($val->name, 4, strpos($val->name, '\\', 5) - 4);
-					return strncmp($val->name, 'yii\\', 4) === 0 && !in_array($subName, $self->extensions);
-				});
-				break;
-			default:
-				$types = array_filter($types, function($val) use ($navClasses) {
-					return strncmp($val->name, "yii\\$navClasses\\", strlen("yii\\$navClasses\\")) === 0;
-				});
-		}
-		return $types;
 	}
 }
